server/pyproject.toml
[project]
name = "agent-router"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "logfire[fastapi]>=4.27.0",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pgvector>=0.4.2",
    "pydantic-ai>=1.67.0",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.8.0",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
